Relicense under LGPL-2.1 to avoid GPL-3.0 binary image license
We're going to use LGPL-2.1 code that I'm unwilling at present to relicense.
We're also loaded dynamically into mate-notification-daemon with dlopen/dlsym,
bringing licenses into question. If we ever chose to link against other libs,
those affect the process space in mate-notification-daemon if we remain as
GPL-2.0.

In all likeliness this would result in the running code being deemed GPL-3.0,
which is unacceptable and infectious in this context. To preserve the true
licensing conditions of Solus project code, we ensure this .so will have
linking exception though LGPL-2.1.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>