engine: Hook up GModule init/destroy functions
When we've fully implemented this theme engine, we'll be dynamically loaded
by the GModule system. This is essentially dlopen/dlsym, but it does also
automatically call the init and destroy functions.

Given that we'll be loading theme resources into the global style context,
we should use this opportunity to then unload them again to ensure there is
no adverse affect on the other themes in the notification daemon process.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>