Add support for action-icons, though the daemon doesn't currently support them
At some point we hope (through patches from ourselves) that the notification
daemon will broadcast the action-icons capability. Thus it makes sense to
have the behaviour already demonstratobly implemented in our theme engine
and be ready for the switch-over.

This also shows as a proof of concept of exactly *how* to implement it.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>