Do not register any dynamic types
Long story short we register dynamic types, and those would require modification
to the MATE Notification Daemon to use GTypeModule. Unfortunately after much
fiddling, and a local conversion of the daemon to do so, it still struggles
as GTypeModule never properly unregisters the interfaces and types, so whenever
we reload the plugin it causes the type to be re-registered, leading to all
kinds of funkiness.

To work around this, and not require patches to the daemon itself, we drop all
use of dynamic types, rely on GtkBuilder, and manage everything with a custom
struct. Technically this route is lighter weight too, so yay.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>